Updating and cleaning the project files
